Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/svg/animations/animate-path-animation-Ll-Vv-Hh-expected.txt


FAIL Test path animation where coordinate modes of start and end differ. You should see PASS messages assert_equals: expected "M -15 -15 L 15 0 V 15 H 0 Z" but got "M -30 -30 L 30 0 V 30 H 0 Z"
